Made adjustments to the file structure incdluing moving migrations and altering the config and __init__ files to more closely match what we have been taught.
