+++ app/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask_migrate import Migrate 
 from flask_sqlalchemy import SQLAlchemy
-from app.config import Config 
+from config import Config 
 from flask_login import LoginManager
 
 app = Flask(__name__)
