+++ app/templates/base.html
@@ -35,10 +35,9 @@
                 <li><a href="/logout">Logout</a></li>
 
             {% else %}
-
-                <li><a href="/login">Login</a></li>
-                <li><a href="/signup">Sign Up</a></li>
-
+                <ul class="nav-links">
+                    <li><a href="/login">Login</a></li>
+                </ul>
             {% endif %}
         </nav>
